requirements_updated.txt
# Django Core
Django==5.2.8
djangorestframework==3.15.2
django-cors-headers==4.3.1
django-filter==24.2

# Database
# mysqlclient==2.2.4  # Comentado por problemas en Windows
pymysql==1.1.0  # Alternativa compatible con Windows

# Channels (WebSockets)
channels==4.0.0
channels-redis==4.2.0
daphne==4.1.0

# Celery (Tareas asíncronas)
celery==5.3.6
redis==5.0.1

# JWT
PyJWT==2.8.0

# HTTP Requests
requests==2.31.0

# API Documentation
drf-spectacular==0.27.1

# Image Processing
# Pillow==10.4.0

# Utilities
python-decouple==3.8

# Testing
pytest==8.0.0
pytest-django==4.7.0
factory-boy==3.3.0
